RavenDB-22835 : remove unused code
