+++ src/Raven.Server/Documents/Handlers/CountersHandler.cs
@@ -7,28 +7,21 @@
 using System;
 using System.Collections.Generic;
 using System.Net;
-using System.Net.Http;
 using System.Threading.Tasks;
 using Raven.Client;
 using Raven.Client.Documents.Changes;
-using Raven.Client.Documents.Conventions;
 using Raven.Client.Documents.Operations.Counters;
 using Raven.Client.Documents.Smuggler;
 using Raven.Client.Exceptions.Documents;
 using Raven.Client.Exceptions.Documents.Counters;
-using Raven.Client.Http;
 using Raven.Client.Json.Serialization;
 using Raven.Server.Routing;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.TrafficWatch;
 using Raven.Server.Utils;
-using Voron.Data.Tables;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
 using Sparrow.Server;
-using Voron;
-using System.Linq;
-using Sparrow.Server.Utils;
 
 namespace Raven.Server.Documents.Handlers
 {
@@ -761,232 +754,6 @@ namespace Raven.Server.Documents.Handlers
         {
             throw new DocumentDoesNotExistException(docId, "Cannot operate on counters of a missing document.");
         }
-
-        public class ExecuteFixCounterGroupsCommand : TransactionOperationsMerger.MergedTransactionCommand
-        {
-            private readonly List<string> _docIds;
-            private readonly DocumentDatabase _database;
-
-
-            public ExecuteFixCounterGroupsCommand(DocumentDatabase database, string docId) : this(database, [docId])
-            {
-            }
-
-            public ExecuteFixCounterGroupsCommand(DocumentDatabase database, List<string> docIdsToFix)
-            {
-                _docIds = docIdsToFix;
-                _database = database;
-            }
-
-            protected override long ExecuteCmd(DocumentsOperationContext context)
-            {
-                var numOfCounterGroupFixed = _database.DocumentsStorage.CountersStorage.FixCountersForDocuments(context, _docIds);
-                return numOfCounterGroupFixed;
-            }
-
-            public override TransactionOperationsMerger.IReplayableCommandDto<TransactionOperationsMerger.MergedTransactionCommand> ToDto<TTransaction>(TransactionOperationContext<TTransaction> context)
-            {
-                throw new NotImplementedException();
-            }
-        }
-
-        [RavenAction("/databases/*/counters/fix-document", "PATCH", AuthorizationStatus.ValidUser, EndpointType.Write)]
-        public async Task FixCounterGroupsForDocument()
-        {
-            var docId = GetStringQueryString("id");
-            await FixCounterGroupsInternal(docId);
-        }
-
-        [RavenAction("/databases/*/counters/fix", "PATCH", AuthorizationStatus.ValidUser, EndpointType.Write)]
-        public async Task FixCounterGroupsForDatabase() => await FixCounterGroupsInternal();
-
-        private async Task FixCounterGroupsInternal(string id = null)
-        {
-            var first = GetBoolValueQueryString("first", required: false) ?? true;
-
-            Task task = id == null
-                ? FixAllCounterGroups()
-                : Database.TxMerger.Enqueue(new ExecuteFixCounterGroupsCommand(Database, id));
-
-            if (first == false)
-            {
-                await task;
-                return;
-            }
-
-            var tasks = new List<Task> { task };
-            var toDispose = new List<IDisposable>();
-            try
-            {
-                using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-                using (context.OpenReadTransaction())
-                {
-                    foreach (var node in ServerStore.GetClusterTopology(context).AllNodes)
-                    {
-                        if (node.Value == Server.WebUrl)
-                            continue;
-
-                        var requestExecutor = ClusterRequestExecutor.CreateForShortTermUse(node.Value, Server.Certificate.Certificate, DocumentConventions.DefaultForServer);
-                        toDispose.Add(requestExecutor);
-
-                        var cmd = new FixCounterGroupsCommand(Database.Name);
-
-                        toDispose.Add(requestExecutor.ContextPool.AllocateOperationContext(out JsonOperationContext ctx));
-                        tasks.Add(requestExecutor.ExecuteAsync(cmd, ctx));
-                    }
-                }
-
-                await Task.WhenAll(tasks);
-            }
-            finally
-            {
-                foreach (var disposable in toDispose)
-                {
-                    disposable.Dispose();
-                }
-            }
-        }
-
-        private async Task FixAllCounterGroups()
-        {
-            const int maxNumberOfDocsToFixInSingleTx = 1024;
-            List<string> docIdsToFix = new();
-            StartAfterSliceHolder startAfterSliceHolder = null;
-            string lastDocId = null;
-
-            try
-            {
-                while (true)
-                {
-                    using (Database.DocumentsStorage.ContextPool.AllocateOperationContext(out DocumentsOperationContext context))
-                    using (context.OpenReadTransaction())
-                    {
-                        var table = new Table(CountersStorage.CountersSchema, context.Transaction.InnerTransaction);
-
-                        while (true)
-                        {
-                            bool hasMore = false;
-
-                            foreach (var (_, tvh) in table.SeekByPrimaryKeyPrefix(Slices.BeforeAllKeys, startAfter: startAfterSliceHolder?.GetStartAfterSlice(context) ?? Slices.Empty, skip: 0))
-                            {
-                                hasMore = true;
-
-                                using (var docId = CountersStorage.ExtractDocId(context, ref tvh.Reader))
-                                {
-                                    if (docId != lastDocId)
-                                    {
-                                        lastDocId = docId;
-
-                                        using (var old = startAfterSliceHolder)
-                                        {
-                                            startAfterSliceHolder = new StartAfterSliceHolder(lastDocId);
-                                        }
-                                    }
-
-                                    using (var data = CountersStorage.GetCounterValuesData(context, ref tvh.Reader))
-                                    {
-                                        data.TryGet(CountersStorage.Values, out BlittableJsonReaderObject counterValues);
-                                        data.TryGet(CountersStorage.CounterNames, out BlittableJsonReaderObject counterNames);
-
-                                        if (counterValues.Count == counterNames.Count)
-                                        {
-                                            var counterValuesPropertyNames = counterValues.GetSortedPropertyNames();
-                                            var counterNamesPropertyNames = counterNames.GetSortedPropertyNames();
-                                            if (counterValuesPropertyNames.SequenceEqual(counterNamesPropertyNames))
-                                                continue;
-                                        }
-                                    }
-
-                                    // Document is corrupted; add to list and break to skip remaining CounterGroups of current document
-                                    docIdsToFix.Add(lastDocId);
-
-                                    break;
-                                }
-                            }
-
-                            if (docIdsToFix.Count == 0)
-                                return;
-
-                            if (docIdsToFix.Count < maxNumberOfDocsToFixInSingleTx && hasMore)
-                                continue;
-
-                            await Database.TxMerger.Enqueue(new ExecuteFixCounterGroupsCommand(Database, docIdsToFix));
-                            docIdsToFix.Clear();
-
-                            if (hasMore)
-                                break;  // break from inner while loop in order to open a new read tx
-
-                            startAfterSliceHolder?.Dispose();
-                            return;
-                        }
-
-                    }
-                }
-            }
-            finally
-            {
-                startAfterSliceHolder?.Dispose();
-            }
-        }
-
-
-        private class StartAfterSliceHolder : IDisposable
-        {
-            private readonly string _docId;
-
-            private readonly List<IDisposable> _toDispose = [];
-
-
-            public StartAfterSliceHolder(string docId)
-            {
-                _docId = docId;
-            }
-
-            public void Dispose()
-            {
-                foreach (var scope in _toDispose)
-                {
-                    scope.Dispose();
-                }
-
-                _toDispose.Clear();
-            }
-
-            public unsafe Slice GetStartAfterSlice(DocumentsOperationContext context)
-            {
-                _toDispose.Add(DocumentIdWorker.GetSliceFromId(context, _docId, out Slice documentKeyPrefix, separator: SpecialChars.RecordSeparator));
-                _toDispose.Add(context.Allocator.Allocate(documentKeyPrefix.Size + sizeof(long), out var startAfterBuffer));
-                _toDispose.Add(Slice.External(context.Allocator, startAfterBuffer.Ptr, startAfterBuffer.Length, out var startAfter));
-
-                documentKeyPrefix.CopyTo(startAfterBuffer.Ptr);
-                *(long*)(startAfterBuffer.Ptr + documentKeyPrefix.Size) = long.MaxValue;
-
-                return startAfter;
-            }
-        }
-
-        private class FixCounterGroupsCommand : RavenCommand
-        {
-            private readonly string _database;
-
-            public FixCounterGroupsCommand(string database)
-            {
-                _database = database;
-            }
-
-            public override bool IsReadRequest => false;
-
-            public override HttpRequestMessage CreateRequest(JsonOperationContext ctx, ServerNode node, out string url)
-            {
-                url = $"{node.Url}/databases/{_database}/counters/fix?first=false";
-
-                return new HttpRequestMessage
-                {
-                    Method = HttpMethod.Patch
-                };
-            }
-        }
-
     }
 
     public class ExecuteCounterBatchCommandDto : TransactionOperationsMerger.IReplayableCommandDto<CountersHandler.ExecuteCounterBatchCommand>
